+++ R/kimisc-package.R
@@ -15,8 +15,8 @@
 #' \tabular{ll}{
 #'   Package: \tab kimisc\cr
 #'   Type: \tab Package\cr
-#'   Version: \tab 0.2\cr
-#'   Date: \tab 2014-03-20\cr
+#'   Version: \tab 0.2-1\cr
+#'   Date: \tab 2014-03-31\cr
 #'   License: \tab GPL v3\cr
 #' }
 NULL